게시글 댓글 ajax 추가
댓글 삭제시 오류 발생

+++ src/main/java/com/example/demo/JPA/Service/BoardReplyService.java
@@ -11,13 +11,12 @@ import com.example.demo.JPA.Repository.BoardRepositoryMany;
 import com.example.demo.JPA.Repository.UserRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
-import java.sql.Date;
+import java.time.LocalDateTime;
 import java.util.List;
 
 @Service
@@ -28,6 +27,7 @@ public class BoardReplyService {
     private final BoardRepositoryMany boardRepositoryMany;
 
     private  final BoardReplyRepository boardReplyRepository;
+    private final UserRepository userRepository;
 
     public boardReadFormDto getBoardWithReply(Long boardManyId, int page, int size) {
         BoardMany boardMany = boardRepositoryMany.findById(boardManyId).orElse(null);
@@ -41,25 +41,53 @@ public class BoardReplyService {
         return null;
     }
 
-    public BoardReply save(BoardReplyDto boardReplyDto) {
+    public boardReadFormDto getBoardWithReply(Long boardManyId){
+        BoardMany boardMany = boardRepositoryMany.findById(boardManyId).orElse(null);
+        if (boardMany != null) {
+            List<BoardReply> boardReplies = boardReplyRepository.findByBoardmanyId(boardManyId);
+            List<BoardReplyDto> boardReplyDtos =new BoardReplyDto().toDtoList(boardReplies);
+            return new boardReadFormDto(boardMany, boardReplyDtos);
+        }
+        return null;
+    }
+
+    public List<BoardReplyDto> getReplies(Long boardManyId){
+        List<BoardReply> boardReplies = boardReplyRepository.findByBoardmanyId(boardManyId);
+        List<BoardReplyDto> boardReplyDtos =new BoardReplyDto().toDtoList(boardReplies);
+        return boardReplyDtos;
+    }
+
+
+    public BoardReply save ( BoardReplyDto boardReplyDto ) {
 
         BoardMany boardMany = boardRepositoryMany.findById(boardReplyDto.getBoardId()).orElse(null);
         if (boardMany == null) {
             return null;
         }
-        User user = boardMany.getUser();
-
+        User user = userRepository.findById(boardReplyDto.getUserId()).orElse(null);
+        if(user==null){
+            return null;
+        }
+        //이 부분 함수로 entity함수로 처리하기
         BoardReply boardReply = new BoardReply();
         boardReply.setReplyContent(boardReplyDto.getReplyContent());
-        boardReply.setCreateDate(new Date(System.currentTimeMillis()));
+        boardReply.setCreateDate(LocalDateTime.now());
         boardReply.setUser(user);
         boardReply.setBoardmany(boardMany);
 
         return boardReplyRepository.save(boardReply);
     }
 
-    public void delete(BoardReply boardReply) {
-        boardReplyRepository.delete(boardReply);
+    public void delete(Long replyId, String userId ) {
+
+
+        BoardReply boardReply = boardReplyRepository.findById(replyId).orElse(null);
+        if (boardReply != null) {
+            if(boardReply.getUser().getId().equals(userId)){
+                boardReplyRepository.delete(boardReply);
+            }
+
+        }
     }
